Remove Tracking Pixel
This seems to be used to ping home, probably done by the author for the purpose of statistics, to know how many projects are using Django Jet, and also to keep some control on who was paying or not paying for the commercial license. I think this ping.gif should be removed.
Aditionally, the server where the image is located is not working, so this is taking 20 secods before the browser blocks the request, which might cause lots of different problems.